server/requirements.txt
# Python 의존성
# Demucs 및 오디오 처리 라이브러리

demucs>=4.0.0
torch>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0
